requirements.txt
streamlit==1.31.0
torch==2.1.0
torchvision==0.16.0
pillow==10.2.0
matplotlib==3.8.0
numpy==1.26.0
